Adding authorization guards

+++ Lab/Exercises/Cubicle-Workshop-Part-2-Resources/Cubicle-Workshop-Part-2/src/controllers/delete.js
@@ -4,23 +4,49 @@ const isUser = require('../middleware/isUserGuard');
 const { findCubeById, deleteCube } = require('../services/cubeService');
 
 deleteController.get('/:cubeId', isUser, async (req, res) => {
-    const cube = await findCubeById(req.params.cubeId);
-    const difficultyMap = {
-        '1': `1 - Very Easy`,
-        '2': `2 - Easy`,
-        '3': `3 - Medium (Standard 3x3)`,
-        '4': `4 - Intermediate`,
-        '5': `5 - Expert`,
-        '6': `6 - Hardcore`
+    try {
+        const cube = await findCubeById(req.params.cubeId);
+
+        const isOwner = req.user.id == cube.creatorId;
+
+        if (isOwner) {
+            const difficultyMap = {
+                '1': `1 - Very Easy`,
+                '2': `2 - Easy`,
+                '3': `3 - Medium (Standard 3x3)`,
+                '4': `4 - Intermediate`,
+                '5': `5 - Expert`,
+                '6': `6 - Hardcore`
+            }
+
+            const difficultyString = difficultyMap[cube.difficulty];
+            return res.render('deleteCube', { cube, difficulty: difficultyString, user: req.user });
+        }
+
+        res.redirect('/');
+    } catch (error) {
+        res.redirect('/404');
     }
 
-    const difficultyString = difficultyMap[cube.difficulty];
-    res.render('deleteCube', { cube, difficulty: difficultyString, user: req.user });
+
 });
 
 deleteController.post('/:cubeId', isUser, async (req, res) => {
-    await deleteCube(req.params.cubeId);
-    res.redirect('/');
+    try {
+        const cube = await findCubeById(req.params.cubeId);
+
+        const isOwner = req.user.id == cube.creatorId;
+
+        if (isOwner) {
+            await deleteCube(req.params.cubeId);
+            return res.redirect('/');
+        }
+
+        res.redirect('/404');
+
+    } catch (error) {
+        res.redirect('/404');
+    }
 });
 
 module.exports = deleteController;
